Committed changes

+++ src/sk/com/ymca/javaStart/lecture8/classwork/ex1/ArithmeticOperations.java
@@ -1,21 +1,19 @@
 package sk.com.ymca.javaStart.lecture8.classwork.ex1;
 
-import java.io.IOException;
-
 public class ArithmeticOperations {
-    public int add(int number1, int number2) {
+    public double add(double number1, double number2) {
         return number1 + number2;
     }
 
-    public int sub(int number1, int number2) {
+    public double sub(double number1, double number2) {
         return number1 - number2;
     }
 
-    public int mul(int number1, int number2) {
+    public double mul(double number1, double number2) {
         return number1 * number2;
     }
 
-    public int div(int number1, int number2) {
+    public double div(double number1, double number2) {
         return number1 / number2;
     }
 }
